+++ rotkehlchen/chain/evm/node_inquirer.py
@@ -36,7 +36,7 @@ from rotkehlchen.chain.evm.constants import (
 )
 from rotkehlchen.chain.evm.contracts import EvmContract, EvmContracts
 from rotkehlchen.chain.evm.proxies_inquirer import EvmProxiesInquirer
-from rotkehlchen.chain.evm.types import NodeName, Web3Node, WeightedNode
+from rotkehlchen.chain.evm.types import NodeName, RemoteDataQueryStatus, Web3Node, WeightedNode
 from rotkehlchen.constants import ONE
 from rotkehlchen.errors.misc import (
     BlockchainQueryError,
@@ -1435,7 +1435,7 @@ class EvmNodeInquirer(ABC, LockableQueryMixIn):
             force_refresh: bool = False,
             chain_id: ChainID | None = None,
             cache_key_parts: Sequence[str] | None = None,
-    ) -> bool:
+    ) -> RemoteDataQueryStatus:
         """
         It checks if the cache data is fresh enough and if not, it queries
         the remote sources of the data and stores it to the globaldb cache tables.
@@ -1457,7 +1457,7 @@ class EvmNodeInquirer(ABC, LockableQueryMixIn):
             force_refresh is False
         ):
             log.debug(f'Not refreshing cache {cache_type}. Queried recently')
-            return False
+            return RemoteDataQueryStatus.NO_UPDATE
 
         try:
             new_data = query_method(
@@ -1469,9 +1469,12 @@ class EvmNodeInquirer(ABC, LockableQueryMixIn):
             log.error(
                 f'Failed to call {query_method} when updating cache {cache_type} due to {e}',
             )
-            return False
+            return RemoteDataQueryStatus.FAILED
 
-        return new_data is not None
+        return (
+            RemoteDataQueryStatus.NEW_DATA if new_data is not None
+            else RemoteDataQueryStatus.NO_UPDATE
+        )
 
 
 class EvmNodeInquirerWithDSProxy(EvmNodeInquirer):